Fixed bug: 0.1.6

+++ Source/UIColor+Extension.swift
@@ -84,9 +84,9 @@ extension UIColor {
      - Returns: A random color object.
      */
     public func random(alpha: CGFloat = 1) -> UIColor {
-        let hue = arc4random() % 256 / 256  //  0.0 to 1.0
-        let saturation = (arc4random() % 128 / 256) + 0.5  //  0.5 to 1.0, away from white
-        let brightness = (arc4random() % 128 / 256) + 0.5  //  0.5 to 1.0, away from black
+        let hue = CGFloat(arc4random() % 256 / 256)  //  0.0 to 1.0
+        let saturation = CGFloat(arc4random() % 128 / 256) + 0.5  //  0.5 to 1.0, away from white
+        let brightness = CGFloat(arc4random() % 128 / 256) + 0.5  //  0.5 to 1.0, away from black
         return UIColor(hue: hue, saturation: saturation, brightness: brightness, alpha: alpha)
     }
